requirements.txt
Pillow>=7.2.0
numpy>=1.18.4
tensorboard>=2.3.0
torch>=1.5.1+cu111
torchvision>=0.6.1+cu111
